Massive refactoring of how headers are handled
Individual headers and now objects, and each attribute of a header is an object.

The ContentType & Disposition classes are now subclasses of Courriel::Header.

The old Courriel::ContentType && ::Disposition classes still exist, but just
die when loaded.

Along the way, fixed parsing of header attributes to handle continuations and encodings.

Still need to fix header attribute output to handle this stuff.